+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+# 1.1.7
+
+### New features
+
+* JDBC driver upgrade (v0.4.1 -> [v0.4.6](https://github.com/ClickHouse/clickhouse-java/releases/tag/v0.4.6))
+* Support DateTime64 by [@lucas-tubi](https://github.com/lucas-tubi) ([#165](https://github.com/ClickHouse/metabase-clickhouse-driver/pull/165))
+* Use native `startsWith`/`endsWith` instead of `LIKE str%`/`LIKE %str`
+
+# 1.1.6
+
+### Bug fixes
+
+* Fixed temporal bucketing issues (see [#155](https://github.com/ClickHouse/metabase-clickhouse-driver/issues/155))
+
+# 1.1.5
+
+### Bug fixes
+
+* Fixed Nippy error on cached questions (see [#147](https://github.com/ClickHouse/metabase-clickhouse-driver/issues/147))
+
 # 1.1.4
 
 ### Bug fixes